feat: Add useCanvasProvider hook and customTheme prop
