+++ modules/react/_examples/stories/mdx/Density.mdx
@@ -1,5 +1,6 @@
 import {ExampleCodeBlock} from '@workday/canvas-kit-docs';
 import {Density} from './examples/Density';
+import {DensityComponents} from './examples/DensityComponents';
 
 <Meta title="Examples/Forms/Density and Alignment" />
 
@@ -10,3 +11,7 @@ options to users, especially in the context of forms. Below is an example of how
 from elements to achieve the desired density.
 
 <ExampleCodeBlock code={Density} />
+
+To view how density is being applied to components, view below.
+
+<ExampleCodeBlock code={DensityComponents} />
